+++ packages/components/psammead-media-player/src/Canonical/index.jsx
@@ -10,14 +10,14 @@ const Canonical = ({ src, title, placeholderSrc }) => {
   `;
 
   const StyledIframe = styled.iframe`
+    border: 0;
+    left: 0;
+    overflow: hidden;
     position: absolute;
     top: 0;
-    left: 0;
     width: 100%;
     height: 100%;
-    border: 0;
-    overflow: hidden;
-    ${placeholderSrc ? backgroundStyle : ''}
+    ${placeholderSrc ? backgroundStyle : null}
   `;
 
   return (