+++ packages/components/psammead-script-link/src/index.stories.jsx
@@ -3,12 +3,14 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import styled from 'styled-components';
 import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
+import { GEL_SPACING_DBL } from '@bbc/gel-foundations/spacings';
+import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
 import notes from '../README.md';
 import ScriptLink from './index';
 
 const Container = styled.div`
   background-color: black;
-  padding: 1rem;
+  padding: ${GEL_SPACING_DBL};
   height: 100vh;
 `;
 
@@ -29,7 +31,8 @@ storiesOf('Components|ScriptLink', module)
             href="https://www.bbc.com/serbian/lat"
             variant={variant}
           >
-            {label}
+            <span>{label}</span>
+            <VisuallyHiddenText>{variant}</VisuallyHiddenText>
           </ScriptLink>
         </Container>
       );